+++ src/index.ts
@@ -4,4 +4,4 @@ export * from './func';
 export * from './object';
 export * from './string';
 export * from './types';
-export * from './subject';
+export * from './subscriber';