+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @xiao-ai/utils
 
+## 1.6.0
+
+feat:
+  - Subject 类更名为 Subscriber
+  - 订阅器参数的“名称”可以用数字类型
+
+## 1.5.2
+
+fix:
+  - 修正 stringifyClass 输入数据类型
+  - 修复 useFollow 会无限触发的 effect 的错误
+
 ## 1.5.1
 
 fix: